+++ pkg/selector/selector_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"testing/fstest"
 
-	"github.com/nao1215/jsonize/internal/registry"
+	"github.com/nao1215/jsonize/pkg/registry"
 )
 
 func buildRegistry(t *testing.T, defs map[string]string) *registry.Registry {